Fix seek error that would find the wrong record number.
The crux of this was the seek would find a valid tail location in a later block
from the writing of longer/larger file. This would locate the precise tail of
the other file and so no further seeking would occur within the block and the
record was being left as 0.

+++ tests/hosted/signed_log_tests.cpp
@@ -63,13 +63,13 @@ TEST_F(SignedLogSuite, AppendingAnEntry) {
 
     auto file_read = storage.file(Storage::Meta);
     ASSERT_TRUE(file_read.seek_end());
-    ASSERT_EQ(file_read.record(), 1);
+    ASSERT_EQ(file_read.record(), (uint32_t)1);
 
     Storage storage_2{ data_memory_, false };
     ASSERT_TRUE(storage_2.begin());
     auto file_read_2 = storage_2.file(Storage::Meta);
     ASSERT_TRUE(file_read_2.seek_end());
-    ASSERT_EQ(file_read_2.record(), 1);
+    ASSERT_EQ(file_read_2.record(), (uint32_t)1);
 }
 
 TEST_F(SignedLogSuite, AppendingTwoLogs) {
